Rename jotai atoms
