+++ ReactNative/atoms/authAtom.ts
@@ -1,4 +1,3 @@
-// @src/store.js
 import { atom } from "jotai";
 
 export const authAtom = atom({
