+++ ReactNative/atoms/configAtom.ts
@@ -1,0 +1,9 @@
+import { atom } from "jotai";
+
+interface IConfigAtom {
+    devMesssage: string
+}
+
+export const configAtom = atom<IConfigAtom>({
+    devMesssage: ''
+});
